+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT
 - PT-11464 - Recognizes basic settings of price management in Magento
 - PT-11465 - Fixed incorrect migration of min and max purchase
+- NTR - Fix query missing table prefix
 
 # 1.1.1
 - NTR - Solves an issue with version compatibility