+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT
+# 1.1.2
 - PT-11464 - Recognizes basic settings of price management in Magento
 - PT-11465 - Fixed incorrect migration of min and max purchase
 - PT-11468 - Fix seo url reader for categories